Optimization for cook codec. Rework sample output to be able to use highly optimized dsp routines. Moved some functions to iram. Speeds up codec by 1.3 MHz on PP5022.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@24815 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/libcook/cook.h
@@ -97,6 +97,6 @@ typedef struct cook {
 
 int cook_decode_init(RMContext *rmctx, COOKContext *q);
 int cook_decode_frame(RMContext *rmctx,COOKContext *q,
-                      int16_t *outbuffer, int *data_size,
+                      int32_t *outbuffer, int *data_size,
                       const uint8_t *inbuffer, int buf_size);
 #endif /*_COOK_H */
